Prepare lsp multicall
