+++ test/SynthsSdk.test.ts
@@ -1,7 +1,6 @@
 import { ethers, network } from "hardhat";
 import { BigNumber } from "ethers";
 import { expect } from "chai";
-// import axios from "axios";
 import Synths from "../src/index";
 import { SynthsAssetsConfig } from "../src/types/assets.t";
 import Asset from "../src/lib/Asset";
